Revert "SCRUM-368: Fix generate matches multiple times for the same day (#11)"
This reverts commit 00bd593a20ff635ae66655e61f2e5e3b108e8858.
